% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{rwimpute}
\alias{rwimpute}
\title{Imputation using Random Walk}
\usage{
rwimpute(data, num_steps = 10L, seed = 42L, restart_prob = 0)
}
\arguments{
\item{data}{A data frame or matrix containing missing values.}

\item{num_steps}{Number of steps for the random walk. Default is 10.
If -1 and restart_prob is set, it is calculated automatically.}

\item{seed}{Integer, random seed for reproducibility. Default is 42.}

\item{restart_prob}{Probability of restarting the random walk. Default is 0.0.}
}
\value{
A data frame or matrix (depending on the input) with imputed values.
}
\description{
Imputes missing values in a data frame or matrix using a
random walk algorithm based on Gower distance similarity.
}
\examples{
data(iris)
iris[1:5, 1] <- NA
imputed_data <- rwimpute(iris)
}
